+++ px-checkout/src/test/java/com/mercadopago/android/px/internal/features/payment_result/PaymentResultPresenterTest.kt
@@ -3,7 +3,10 @@ package com.mercadopago.android.px.internal.features.payment_result
 import com.mercadopago.android.px.configuration.AdvancedConfiguration
 import com.mercadopago.android.px.internal.repository.PaymentSettingRepository
 import com.mercadopago.android.px.internal.viewmodel.PaymentModel
+import com.mercadopago.android.px.model.PaymentData
+import com.mercadopago.android.px.model.PaymentMethod
 import com.mercadopago.android.px.model.PaymentResult
+import com.mercadopago.android.px.tracking.internal.BankInfoHelper
 import com.mercadopago.android.px.tracking.internal.MPTracker
 import com.mercadopago.android.px.tracking.internal.views.ResultViewTrack
 import org.junit.Before
@@ -12,6 +15,8 @@ import org.junit.runner.RunWith
 import org.mockito.Mock
 import org.mockito.junit.MockitoJUnitRunner
 import org.mockito.kotlin.*
+
+private const val PAYMENT_METHOD_ID = "123456"
 
 @RunWith(MockitoJUnitRunner::class)
 class PaymentResultPresenterTest {
@@ -27,14 +32,21 @@ class PaymentResultPresenterTest {
         val advancedConfiguration = mock<AdvancedConfiguration>()
         val paymentModel = mock<PaymentModel>()
         val paymentResult = mock<PaymentResult>()
+        val paymentData = mock<PaymentData>()
+        val paymentMethod  = mock<PaymentMethod>()
+        val bankInfoHelper = mock<BankInfoHelper>()
+
+        whenever(paymentMethod.id).thenReturn(PAYMENT_METHOD_ID)
+        whenever(paymentData.paymentMethod).thenReturn(paymentMethod)
         whenever(paymentModel.paymentResult).thenReturn(paymentResult)
         whenever(paymentModel.congratsResponse).thenReturn(mock())
         whenever(paymentModel.remedies).thenReturn(mock())
-        whenever(paymentResult.paymentData).thenReturn(mock())
+        whenever(paymentResult.paymentData).thenReturn(paymentData)
         whenever(paymentsSettings.checkoutPreference).thenReturn(mock())
         whenever(paymentsSettings.currency).thenReturn(mock())
         whenever(paymentsSettings.advancedConfiguration).thenReturn(advancedConfiguration)
         whenever(advancedConfiguration.paymentResultScreenConfiguration).thenReturn(mock())
+
         presenter = PaymentResultPresenter(
             paymentsSettings,
             paymentModel,
@@ -42,6 +54,7 @@ class PaymentResultPresenterTest {
             true,
             mock(),
             mock(),
+            bankInfoHelper,
             tracker
         )
     }