+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-## Unreleased
-_XX_03_2022_
+##  VERSION 4.109.0
+_31_03_2022_
 * FIX - Parameter specified as non-null is null - parameter state
+* FIX - Crash when PaymentTypeId is null
+* FEATURE - Remove label "interest free" from the front of the installment amount
+* FEATURE - Added tracks for debin transfer.
 
 ## VERSION 4.108.1
 _24_03_2022_